Move Opening Hours into components and add test
