+++ .jest/setEnvVars.js
@@ -1,3 +1,4 @@
 // Set environment variables for Jest test environment
 
 process.env.COUNCIL_WEBSITE_HOMEPAGE_URL = "https://www.newark-sherwooddc.gov.uk";
+process.env.CUSTOMER_SERVICES_OPENING_TIMES_FULL_DESCRIPTION = "{\"TestDay\":\"12am - 12pm\"}";
